\name{arpproj2-package}
\alias{arpproj2-package}
\alias{arpproj2}
\docType{package}
\title{
\packageTitle{arpproj2}
}
\description{
\packageDescription{arpproj2}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{arpproj2}
\packageIndices{arpproj2}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{arpproj2}

Maintainer: \packageMaintainer{arpproj2}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
# simple examples of the most important functions
}
