% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/f_vector_checker.R
\name{f_vector_checker}
\alias{f_vector_checker}
\title{Check if input data has correct vector form}
\usage{
f_vector_checker(input_vector)
}
\arguments{
\item{input_vector}{A vector.}
}
\value{
Stop if input data is not a vector.
}
\description{
Check if input data has correct vector form
}
\examples{
> f_vector_checker(c(1, 2, 3))
output will be nothing

> f_vector_checker()
Error in stopifnot(is.numeric(input_vector)) ...: argument "input_vector" is missing, with no default

}
