% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sha1_hash.R
\name{get_sha1_hash}
\alias{get_sha1_hash}
\title{Get SHA1 hash for object}
\usage{
get_sha1_hash(input_object = "")
}
\arguments{
\item{object}{A object.}
}
\value{
A SHA1 hash of the input object.
}
\description{
Get SHA1 hash for object
}
\examples{
get_sha1_hash("sdjs")
}
