% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_data.R
\name{load_data}
\alias{load_data}
\title{Load dataset from txt file in "data" subdirectory}
\usage{
load_data(data_name = "")
}
\arguments{
\item{data_name}{A name of the dataset.}
}
\value{
zbior_data A loaded dataset.
}
\description{
Load dataset from txt file in "data" subdirectory
}
\examples{
savings_data <- load_data(data_name = 'savings.txt')
}
