% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_data.R
\name{plot_data}
\alias{plot_data}
\title{Plot data in prepared form}
\usage{
plot_data(our_data)
}
\arguments{
\item{our_data}{a dataset whcih we want to plot.}
}
\value{
Plot dataset in separate window.
}
\description{
Plot data in prepared form
}
\examples{
savings_data <- load_data(data_name = 'savings.txt')
plot_data(savings_data)
}
