% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/savings_country.R
\docType{data}
\name{savings_country}
\alias{savings_country}
\title{The data in the savings_country file concerns the economic situation of residents of 50 countries.}
\format{
A data frame with 50 rows and 6 variables

\describe{
  \item{Country}{name of the country}
  \item{Pop15}{percentage of citizens aged less than 15 years}
  \item{Pop75}{percentage of citizens aged over than 75 years}
  \item{dpi}{net income per capita}
  \item{ddpi}{income growth rate (in \%)}
  \item{Savings}{total savings per person divided by net income}
}
}
\source{
Lectures on statistics at Warsaw University of Technology
}
\usage{
data (savings_country)
}
\description{
The data in the savings_country file concerns the economic situation of residents of 50 countries.
}
\examples{
data (savings_country)
}
\keyword{datasets}
\keyword{savings_country}
